Move hooks and staging code into gitbutler-repo
