+++ crates/gitbutler-branch-actions/src/virtual.rs
@@ -12,14 +12,12 @@ use crate::{
 };
 use anyhow::{anyhow, bail, Context, Result};
 use bstr::{BString, ByteSlice};
-use git2::{ApplyLocation, ApplyOptions, Repository, ResetType};
-use git2_hooks::HookResult;
 use gitbutler_branch::BranchUpdateRequest;
 use gitbutler_branch::{dedup, dedup_fmt};
 use gitbutler_cherry_pick::RepositoryExt as _;
 use gitbutler_command_context::CommandContext;
 use gitbutler_commit::{commit_ext::CommitExt, commit_headers::HasCommitHeaders};
-use gitbutler_diff::{trees, ChangeType, GitHunk, Hunk};
+use gitbutler_diff::{trees, GitHunk, Hunk};
 use gitbutler_error::error::Code;
 use gitbutler_hunk_dependency::RangeCalculationError;
 use gitbutler_operating_modes::assure_open_workspace_mode;
@@ -29,9 +27,10 @@ use gitbutler_oxidize::{
 use gitbutler_project::access::WorktreeWritePermission;
 use gitbutler_reference::{normalize_branch_name, Refname, RemoteRefname};
 use gitbutler_repo::{
+    hooks,
     logging::{LogUntil, RepositoryExt as _},
     rebase::{cherry_rebase, cherry_rebase_group},
-    RepositoryExt,
+    staging, RepositoryExt,
 };
 use gitbutler_repo_actions::RepoActionsExt;
 use gitbutler_stack::{
@@ -41,7 +40,6 @@ use gitbutler_stack::{
 use gitbutler_time::time::now_since_unix_epoch_ms;
 use itertools::Itertools;
 use serde::Serialize;
-use std::borrow::Cow;
 use std::{collections::HashMap, path::PathBuf, vec};
 use tracing::instrument;
 
@@ -733,65 +731,55 @@ pub fn commit(
     ownership: Option<&BranchOwnershipClaims>,
     run_hooks: bool,
 ) -> Result<git2::Oid> {
-    // get the files to commit
     let diffs = gitbutler_diff::workdir(ctx.repo(), get_workspace_head(ctx)?)?;
-    let statuses = get_applied_status_cached(ctx, None, &diffs)
-        .context("failed to get status by branch")?
-        .branches;
 
-    let (ref mut branch, files) = statuses
-        .into_iter()
-        .find(|(stack, _)| stack.id == stack_id)
-        .with_context(|| format!("stack {stack_id} not found"))?;
+    if ownership.is_none() {
+        // If we are committing everything owned by a branch then we
+        // need to make sure we have first updated ownerships.
+        get_applied_status_cached(ctx, None, &diffs)?;
+    }
 
-    let selected_files = if let Some(ownership) = ownership {
-        files
-            .clone()
-            .into_iter()
-            .filter_map(|file| {
-                let hunks = file
+    let mut stack = ctx.project().virtual_branches().get_stack(stack_id)?;
+    let ownership = ownership
+        .map::<Result<&BranchOwnershipClaims>, _>(Ok)
+        .unwrap_or_else(|| Ok(&stack.ownership))?;
+
+    let selected_files = ownership
+        .claims
+        .iter()
+        .map(|claim| {
+            if let Some(diff) = diffs.get(&claim.file_path) {
+                let hunks = claim
                     .hunks
-                    .into_iter()
-                    .filter(|hunk| {
-                        let hunk: GitHunk = hunk.clone().into();
-                        ownership
-                            .claims
+                    .iter()
+                    .filter_map(|claimed_hunk| {
+                        diff.hunks
                             .iter()
-                            .find(|f| f.file_path.eq(&file.path))
-                            .map_or(false, |f| {
-                                f.hunks.iter().any(|h| {
-                                    h.start == hunk.new_start
-                                        && h.end == hunk.new_start + hunk.new_lines
-                                })
+                            .find(|diff_hunk| {
+                                claimed_hunk.start == diff_hunk.new_start
+                                    && claimed_hunk.end == diff_hunk.new_start + diff_hunk.new_lines
                             })
+                            .cloned()
                     })
-                    .collect::<Vec<_>>();
-                if !hunks.is_empty() {
-                    Some((file.path, hunks))
-                } else {
-                    None
-                }
-            })
-            .collect::<Vec<(PathBuf, Vec<VirtualBranchHunk>)>>()
-    } else {
-        files
-            .clone()
-            .into_iter()
-            .map(|file| (file.path, file.hunks))
-            .collect::<Vec<(PathBuf, Vec<VirtualBranchHunk>)>>()
-    };
+                    .collect_vec();
+                Ok((claim.file_path.clone(), hunks))
+            } else {
+                Err(anyhow!("Claim not found in workspace diff"))
+            }
+        })
+        .collect::<Result<Vec<(_, Vec<_>)>>>()?;
 
     let final_message = if run_hooks {
-        run_message_hook(ctx, message.to_owned())?
+        hooks::message(ctx, message.to_owned())?
     } else {
         message.to_owned()
     };
 
     if run_hooks {
-        stage_files(ctx, &selected_files)?;
-        let result = run_pre_commit_hook(ctx);
+        staging::stage_files(ctx, &selected_files)?;
+        let result = hooks::pre_commit(ctx);
         if result.is_err() {
-            unstage_all(ctx)?;
+            staging::unstage_all(ctx)?;
             result?;
         }
     }
@@ -803,10 +791,10 @@ pub fn commit(
 
     let git_repository = ctx.repo();
     let parent_commit = git_repository
-        .find_commit(branch.head())
-        .context(format!("failed to find commit {:?}", branch.head()))?;
+        .find_commit(stack.head())
+        .context(format!("failed to find commit {:?}", stack.head()))?;
 
-    let tree_oid = gitbutler_diff::write::hunks_onto_commit(ctx, branch.head(), selected_files)?;
+    let tree_oid = gitbutler_diff::write::hunks_onto_commit(ctx, stack.head(), selected_files)?;
     let tree = git_repository
         .find_tree(tree_oid)
         .context(format!("failed to find tree {:?}", tree_oid))?;
@@ -836,116 +824,16 @@ pub fn commit(
     };
 
     if run_hooks {
-        git2_hooks::hooks_post_commit(ctx.repo(), Some(&["../.husky"]))
-            .context("failed to run hook")
-            .context(Code::CommitHookFailed)?;
+        hooks::post_commit(ctx)?;
     }
 
     let vb_state = ctx.project().virtual_branches();
-    branch.set_stack_head(ctx, commit_oid, Some(tree_oid))?;
+    stack.set_stack_head(ctx, commit_oid, Some(tree_oid))?;
 
     crate::integration::update_workspace_commit(&vb_state, ctx)
         .context("failed to update gitbutler workspace")?;
 
     Ok(commit_oid)
-}
-
-fn join_output<'a>(stdout: &'a str, stderr: &'a str) -> Cow<'a, str> {
-    let stdout = stdout.trim();
-    if stdout.is_empty() {
-        stderr.trim().into()
-    } else {
-        stdout.into()
-    }
-}
-
-fn run_message_hook(ctx: &CommandContext, mut message: String) -> Result<String> {
-    let hook_result = git2_hooks::hooks_commit_msg(ctx.repo(), Some(&["../.husky"]), &mut message)
-        .context("failed to run hook")
-        .context(Code::CommitHookFailed)?;
-
-    if let HookResult::RunNotSuccessful { stdout, stderr, .. } = &hook_result {
-        return Err(
-            anyhow!("commit-msg hook rejected: {}", join_output(stdout, stderr))
-                .context(Code::CommitHookFailed),
-        );
-    }
-    Ok(message)
-}
-
-fn run_pre_commit_hook(ctx: &CommandContext) -> Result<()> {
-    let hook_result = git2_hooks::hooks_pre_commit(ctx.repo(), Some(&["../.husky"]))
-        .context("failed to run hook")
-        .context(Code::CommitHookFailed)?;
-
-    if let HookResult::RunNotSuccessful { stdout, stderr, .. } = &hook_result {
-        return Err(
-            anyhow!("commit hook rejected: {}", join_output(stdout, stderr))
-                .context(Code::CommitHookFailed),
-        );
-    }
-    Ok(())
-}
-
-pub(crate) fn unstage_all(ctx: &CommandContext) -> Result<()> {
-    let repo = ctx.repo();
-    // Get the HEAD commit (current commit)
-    let head_commit = repo.head()?.peel_to_commit()?;
-    // Reset the index to match the HEAD commit
-    repo.reset(head_commit.as_object(), ResetType::Mixed, None)?;
-    Ok(())
-}
-
-fn diff_workdir_to_index<'a>(repo: &'a Repository, path: &PathBuf) -> Result<git2::Diff<'a>> {
-    let index = repo.index()?;
-    let mut diff_opts = git2::DiffOptions::new();
-    diff_opts
-        .recurse_untracked_dirs(true)
-        .include_untracked(true)
-        .show_binary(true)
-        .show_untracked_content(true)
-        .ignore_submodules(true)
-        .context_lines(3)
-        .pathspec(path);
-    Ok(repo.diff_index_to_workdir(Some(&index), Some(&mut diff_opts))?)
-}
-
-fn stage_file(ctx: &CommandContext, path: &PathBuf, hunks: &Vec<VirtualBranchHunk>) -> Result<()> {
-    let repo = ctx.repo();
-    let mut index = repo.index()?;
-    if hunks.iter().any(|h| h.change_type == ChangeType::Untracked) {
-        index.add_path(path)?;
-        index.write()?;
-        return Ok(());
-    }
-
-    let mut apply_opts = ApplyOptions::new();
-    apply_opts.hunk_callback(|cb_hunk| {
-        cb_hunk.map_or(false, |cb_hunk| {
-            for hunk in hunks {
-                if hunk.start == cb_hunk.new_start()
-                    && hunk.end == cb_hunk.new_start() + cb_hunk.new_lines()
-                {
-                    return true;
-                }
-            }
-            false
-        })
-    });
-
-    let diff = diff_workdir_to_index(repo, path)?;
-    repo.apply(&diff, ApplyLocation::Index, Some(&mut apply_opts))?;
-    Ok(())
-}
-
-fn stage_files(
-    ctx: &CommandContext,
-    files_to_stage: &Vec<(PathBuf, Vec<VirtualBranchHunk>)>,
-) -> Result<()> {
-    for (path_to_stage, hunks_to_stage) in files_to_stage {
-        stage_file(ctx, path_to_stage, hunks_to_stage)?;
-    }
-    Ok(())
 }
 
 pub(crate) fn push(
